perf(HWSplitSeqComb): first collect and then modify
Calling of `insertInput` and `insertOutput` is slow. Previously we
called them every time a new port should be created. In this
commit, we first collect all ports should be inserted, and then
insert them using `appendInputs` and `appendOutputs` in one go.
This significantly improve the performance, from ~34s to ~0.5s,
which is ~68x speedup.